Match type of process_packet's returned bcnt value
The process_packet() function in hera_catcher_ibvpkt_thread.c returns a
bcnt value as a `uint32_t`.  Returning `0xffffffff` and storing it in a
uint64_t variable results in the value `0x00000000ffffffff`, which does
not compare equal to `-1`!  This resulted in locking/updating/unlocking
the status buffer on every packet resulting in poor throughput.

+++ src/hera_catcher_ibvpkt_thread.c
@@ -831,7 +831,7 @@ static void *run(hashpipe_thread_args_t * args)
         packet_count++;
 
         // Copy packet into any blocks where it belongs.
-        const uint64_t bcnt = process_packet((hera_catcher_bda_input_databuf_t *)db, p_frame);
+        const uint32_t bcnt = process_packet((hera_catcher_bda_input_databuf_t *)db, p_frame);
 
         clock_gettime(CLOCK_MONOTONIC, &stop);
         wait_ns = ELAPSED_NS(recv_start, start);
